Change block to a vector
The object block used to be a fixed, compile-time size based on the
constant BlockSize. The problem with this was when we wanted to encrypt
some data, the resulting ciphertext may be bigger than the plaintext
(especially if we include a random IV at the start).

Now block is a vector of bytes, it can be resized.