Block stores can detect if they were serialised
BlockStore now contains the virtual function WasSerialised(). This
function returns true if the data has been stored from a previous run,
and false otherwise.

This lets ORAM know if the tree data needs be be intialised, or if it
can just be used.

Note that for the FileStore, WasSerialised() returns true if the file
length is the same as the expected length. No integrity checks are made
on the data!