Refactor ORAM to use BlockStore
ORAM now works with dynamic-sized blocks of data, instead of
fixed-sized blocks that are set at compile-time. It means that the size
of encrypted blocks doesn't have to be manually calculated: it can be
calculated at runtime.

The main benefit of this change is that ORAM no longer relies on the
BucketTree class; it can use any give BlockStore to store the Path ORAM
tree. It gives ORAM the flexibility to use a RAM-backed tree, a
file-backed tree, and oppertunity for more block stores to be used in
the future.

ORAM is now also a BlockDevice itself! This means that we can perform
operations on the Agent, plain files, RAM, and now Path ORAM using the
same interface. For instance, all 4 of those BlockDevices can be used
by FileSystem to add and remove files that are backed on those devices.