Modify FileSystem to work on BlockStore
The FileSytem can now work on any block device that uses the BlockStore
interface. Previously it was just limited to working with ORAM, so you
weren't able to add/remove files using the Agent.

You now not only can add/remove files from the Agent, but you can also
do the same to an unencrypted store such as FileStore or RAMStore for
comparison.